# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student Follow-up API",
        "description": "API documentation for the Student Follow-up System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get all students",
                "operationId": "5a54d6df45ca455ad022d2f7b9da9731",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "operationId": "aa0044bcc7e0cbaa736f0c20a4a6765c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "class_id",
                                    "trainer_id",
                                    "name",
                                    "priority",
                                    "status"
                                ],
                                "properties": {
                                    "class_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "trainer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "follow_up_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-01",
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "Medium",
                                        "enum": [
                                            "Low",
                                            "Medium",
                                            "High"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Pending",
                                        "enum": [
                                            "Pending",
                                            "In Progress",
                                            "Completed",
                                            "Archived"
                                        ]
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get a single student",
                "operationId": "5e2013c1980b929b399fa6e9296534b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update a student",
                "operationId": "35f77190e85ebd1a750bc9a367dfe651",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "class_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "trainer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "follow_up_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "Low",
                                            "Medium",
                                            "High"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "In Progress",
                                            "Completed",
                                            "Archived"
                                        ]
                                    },
                                    "position": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete a student",
                "operationId": "0a7e07013ebd062ee8976492953e0d2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Student deleted"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "class_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "trainer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "photo": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "follow_up_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-08-01",
                        "nullable": true
                    },
                    "priority": {
                        "type": "string",
                        "example": "Medium",
                        "enum": [
                            "Low",
                            "Medium",
                            "High"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "Pending",
                        "enum": [
                            "Pending",
                            "In Progress",
                            "Completed",
                            "Archived"
                        ]
                    },
                    "position": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "Students"
        }
    ]
}